Extracted model
